feat(CB2-15349): better way of doing tests
